requirements.txt
discord.py[voice]==2.2.2
python-dotenv==0.21.1
